CRUD для торговых представителей

+++ database/migrations/2024_11_08_135218_create_channels_table.php
@@ -24,6 +24,8 @@ return new class extends Migration {
     public function down()
     : void
     {
-        Schema::dropIfExists('channels');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('channels');
+        }
     }
 };
